#224 Add Participant Class to hold User's Schedule Event Participation Details

+++ src/main/java/io/github/externschool/planner/service/UserServiceImpl.java
@@ -1,11 +1,13 @@
 package io.github.externschool.planner.service;
 
 import io.github.externschool.planner.dto.UserDTO;
+import io.github.externschool.planner.entity.Participant;
 import io.github.externschool.planner.entity.User;
 import io.github.externschool.planner.entity.VerificationKey;
 import io.github.externschool.planner.entity.profile.Person;
 import io.github.externschool.planner.entity.profile.Student;
 import io.github.externschool.planner.entity.profile.Teacher;
+import io.github.externschool.planner.entity.schedule.ScheduleEvent;
 import io.github.externschool.planner.exceptions.EmailExistsException;
 import io.github.externschool.planner.exceptions.RoleNotFoundException;
 import io.github.externschool.planner.repository.UserRepository;
@@ -19,9 +21,9 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
 
 @Service
+@Transactional
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final RoleService roleService;
@@ -52,20 +54,22 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User saveOrUpdate(User user) {
+    public User save(User user) {
         return userRepository.save(user);
     }
 
-    @Transactional
     @Override
     public void deleteUser(final User user) {
-        if (user != null) {
+        if (user != null && findUserByEmail(user.getEmail()) != null) {
             if (user.getVerificationKey() != null) {
                 user.removeVerificationKey();
             }
-            scheduleService.getEventsByOwner(user).stream()
-                    .filter(Objects::nonNull)
-                    .forEach(e -> scheduleService.deleteEvent(e.getId()));
+            for (ScheduleEvent event : user.getOwnEvents()) {
+                scheduleService.deleteEventById(event.getId());
+            }
+            for (Participant participant : user.getParticipants()) {
+                scheduleService.removeParticipant(participant);
+            }
 
             userRepository.delete(user);
         }
@@ -140,7 +144,6 @@ public class UserServiceImpl implements UserService {
         return user;
     }
 
-    @Transactional
     @Override
     public void createAndAddNewKeyAndPerson(User user) {
         //TODO Remove key and person saves when Cascade fixed
@@ -150,7 +153,7 @@ public class UserServiceImpl implements UserService {
         person.addVerificationKey(key);
         personRepository.save(person);
         user.addVerificationKey(key);
-        saveOrUpdate(user);
+        save(user);
         assignNewRolesByKey(user, key);
     }
